+++ src/describe/legacy.rs
@@ -380,168 +380,6 @@ fn node_access_url(v: &Value) -> Cow<str> {
     }
 }
 
-/// Utility function for describe to print service info
-pub fn describe_format_service(v: Value, endpoint_val: Option<Value>) -> String {
-    let port_str = get_ports_str(v.pointer("/spec/ports"), endpoint_val);
-    let fields = vec![
-        (
-            "Name:\t\t",
-            DescItem::MetadataValStr {
-                path: "/name",
-                default: "<No Name>",
-            },
-        ),
-        (
-            "Labels:\t",
-            DescItem::KeyValStr {
-                parent: "/metadata/labels",
-                secret_vals: false,
-            },
-        ),
-        (
-            "Annotations:",
-            DescItem::KeyValStr {
-                parent: "/metadata/annotations",
-                secret_vals: false,
-            },
-        ),
-        ("Created at:\t", DescItem::ObjectCreated),
-        (
-            "Selector:",
-            DescItem::KeyValStr {
-                parent: "/spec/selector",
-                secret_vals: false,
-            },
-        ),
-        (
-            "Type:\t\t",
-            DescItem::ValStr {
-                path: "/spec/type",
-                default: "<No Type>",
-            },
-        ),
-        (
-            "IP:\t\t",
-            DescItem::ValStr {
-                path: "/spec/clusterIP",
-                default: "<No Type>",
-            },
-        ),
-        (
-            "LoadBalIngress:\t",
-            DescItem::ValStr {
-                path: "/status/loadBalancer/ingress/0/hostname",
-                default: "<No Ingress>",
-            },
-        ),
-        ("Ports:\n", DescItem::StaticStr(port_str)),
-        (
-            "SessionAffnity:\t",
-            DescItem::ValStr {
-                path: "/spec/sessionAffnity",
-                default: "<none>",
-            },
-        ),
-    ];
-    describe_object(&v, fields.into_iter())
-}
-
-/// Get ports info out of ports array
-fn get_ports_str(v: Option<&Value>, endpoint_val: Option<Value>) -> Cow<str> {
-    if v.is_none() {
-        return "<none>".into();
-    }
-    let mut buf = String::new();
-    match v.unwrap().as_array() {
-        // safe unwrap, checked above
-        Some(port_array) => {
-            for port in port_array.iter() {
-                let proto = val_str("/protocol", port, "<Unknown>");
-                let name = val_str("/name", port, "<No Name>");
-                let port_num = val_u64("/port", port, 0);
-                let endpoints = match endpoint_val {
-                    Some(ref ep) => {
-                        // to get all the endpoints, we need to check all subsets this port is in
-                        // TODO: This is complex, simplify and/or abstract
-                        let mut epbuf = String::from_str("  Endpoints:\t").unwrap();
-                        let mut found_one = false;
-                        ep.pointer("/subsets").map(|s| {
-                            s.as_array().map(|subsets| {
-                                for subset in subsets.iter() {
-                                    // see if this subset has this port by checking if any port in
-                                    // the ports array has the same port number
-                                    let contains = subset
-                                        .pointer("/ports")
-                                        .map(|p| {
-                                            p.as_array()
-                                                .map(|ports_array| {
-                                                    let mut c = false;
-                                                    for port in ports_array.iter() {
-                                                        if port_num == val_u64("/port", port, 0) {
-                                                            c = true;
-                                                        }
-                                                    }
-                                                    c
-                                                })
-                                                .unwrap_or(false)
-                                        })
-                                        .unwrap_or(false);
-                                    if contains {
-                                        // we do have this port, need to add all addresses as
-                                        // endpoints
-                                        found_one = true;
-                                        let port_num = val_u64("/targetPort", port, 0);
-                                        subset.pointer("/addresses").map(|a| {
-                                            a.as_array().map(|addr_array| {
-                                                let mut first = true;
-                                                for addr in addr_array.iter() {
-                                                    if first {
-                                                        first = false;
-                                                    } else {
-                                                        epbuf.push_str(", ");
-                                                    }
-                                                    epbuf.push_str(
-                                                        format!(
-                                                            "{}:{}",
-                                                            val_str("/ip", addr, "<No IP>"),
-                                                            port_num
-                                                        )
-                                                        .as_str(),
-                                                    );
-                                                }
-                                            })
-                                        });
-                                    }
-                                }
-                            })
-                        });
-                        if !found_one {
-                            epbuf.push_str("<none>");
-                        }
-                        epbuf
-                    }
-                    None => "<No Endpoints>".to_owned(),
-                };
-                buf.push_str(format!("  Port:\t\t{} {}/{}\n", name, port_num, proto).as_str());
-                buf.push_str(
-                    format!(
-                        "  NodePort:\t{} {}/{}\n",
-                        val_str("/name", port, "<No Name>"),
-                        val_u64("/nodePort", port, 0),
-                        proto
-                    )
-                    .as_str(),
-                );
-                buf.push_str(endpoints.as_str());
-                buf.push('\n');
-                buf.push('\n');
-            }
-        }
-        None => buf.push_str("<none>"),
-    }
-    buf.into()
-}
-
 /// Utility function to describe a secret
 pub fn describe_format_secret(v: Value) -> String {
     let fields = vec![
